SDC-727 __init__.py excluded from report to not lower coverage
