+++ sonar-project.properties
@@ -10,7 +10,7 @@ sonar.links.issue=https://github.com/usdot-jpo-sdc/sdc-dot-cvp-ingest/issues
 # Language
 sonar.language=py
 sonar.sources=lambdas, tests
-sonar.exclusions=*coverage*.xml
+sonar.exclusions=*coverage*.xml, **/__init__.py
 sonar.python.coverage.reportPaths=*coverage*.xml
 # =====================================================
 #   Properties that will be shared amongst all modules
